Move `KeyingMaterial` out of ossl module
